refactor: modularize core and processor crates
- split locale.rs into locale/ directory
- moved legacy CSL 1.0 types to legacy.rs
- split options.rs into options/ directory
- split embedded.rs into embedded/ directory by style
- split render.rs into render/ directory
- split csln_analyze/src/main.rs into analyzer and ranker modules
- maintained public API through re-exports in lib.rs
